removing unused dependencies
