+++ apps/web/Cargo.toml
@@ -13,7 +13,6 @@ markdown = "1.0.0-alpha.7"
 reqwest = { version = "0.11", features = ["json", "stream"] }
 serde = "1.0"
 serde_json = "1.0"
-serde-wasm-bindgen = "0.5"
 strum = "0.24"
 strum_macros = "0.24"
 thiserror = "1.0"
