+++ crates/spyglass-plugin/Cargo.toml
@@ -15,7 +15,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 ron = "0.8"
 url = "2.2"
-uuid = { version = "1.0.0", features = ["serde", "v4"], default-features = false }
 
 [lib]
 name = "spyglass_plugin"
