+++ crates/spyglass-searcher/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
-dashmap = "5.2"
 log = "0.4"
 serde = "1.0"
 serde_json = "1.0"
