+++ crates/spyglass/Cargo.toml
@@ -40,7 +40,6 @@ regex = "1"
 reqwest = { version = "0.11", features = ["stream", "json"] }
 ron = "0.8"
 rubato = "0.12.0"
-rusqlite = { version = "*", features = ["bundled"] }
 sentry = "0.30.0"
 sentry-tracing = "0.30.0"
 serde = { version = "1.0", features = ["derive"] }
